+++ src/components/Common/PostCard/PostSaveButton.tsx
@@ -2,6 +2,7 @@ import React, { useCallback } from 'react';
 import AnimatedIconButton from '@components/Common/Button/AnimatedIconButton';
 import { Bookmark } from '@tamagui/lucide-icons';
 import { usePostSaved } from '@src/state';
+import instance from '@src/Instance';
 
 interface IProps {
   itemId: number;
@@ -10,7 +11,9 @@ interface IProps {
 function PostSaveButton({ itemId }: IProps): React.JSX.Element {
   const postSaved = usePostSaved(itemId);
 
-  const doSavePost = useCallback(() => {}, [itemId, postSaved]);
+  const doSavePost = useCallback(() => {
+    void instance.savePost(itemId);
+  }, [itemId, postSaved]);
 
   return (
     <AnimatedIconButton