+++ src/helpers/comments/getParentId.ts
@@ -1,0 +1,5 @@
+export const getParentId = (path: string): number => {
+  const pathArr = path.split('.');
+
+  return Number(pathArr[pathArr.length - 2]);
+};