+++ src/helpers/comments/index.ts
@@ -1,2 +1,3 @@
 export * from './buildCommentChains';
 export * from './createCommentReplyComponents';
+export * from './getParentId';