+++ src/components/NewPost/hooks/useNewPostScreen.tsx
@@ -16,11 +16,17 @@ import {
   useCommunityDefaultLanguage,
   useCurrentAccount,
   useDefaultLanguage,
+  usePostBody,
+  usePostLanguageId,
+  usePostLink,
+  usePostNsfw,
+  usePostTitle,
   useSiteDefaultLanguage,
 } from '@src/state';
 import HeaderButton from '@components/Common/Button/HeaderButton';
 import { PostResponse } from 'lemmy-js-client';
 import instance from '@src/Instance';
+import { updatePost } from '@src/state/post/actions/updatePost';
 
 interface UseNewPostScreen {
   text: string;
@@ -29,7 +35,10 @@ interface UseNewPostScreen {
   nsfw: boolean;
   setNsfw: React.Dispatch<React.SetStateAction<boolean>>;
 
+  title: string;
   setTitle: React.Dispatch<React.SetStateAction<string>>;
+
+  url: string;
   setUrl: React.Dispatch<React.SetStateAction<string>>;
 
   languageId: number;
@@ -51,22 +60,30 @@ export const useNewPostScreen = (isEdit = false): UseNewPostScreen => {
 
   const { communityId, postId } = route.params;
 
+  // Get the possible language options
   const userLanguage = useDefaultLanguage();
   const communityLanguage = useCommunityDefaultLanguage(communityId);
   const siteLanguage = useSiteDefaultLanguage();
 
   const account = useCurrentAccount();
 
+  // We will get the current data if it exists
+  const postContent = postId != null ? usePostBody(postId) : undefined;
+  const postTitle = postId != null ? usePostTitle(postId) : undefined;
+  const postNsfw = postId != null ? usePostNsfw(postId) : undefined;
+  const postLink = postId != null ? usePostLink(postId) : undefined;
+  const postLanguage = postId != null ? usePostLanguageId(postId) : undefined;
+
   const { isLoading, refresh: submit } = useLoadData<
     PostResponse | undefined
   >();
 
-  const [text, setText] = useState('');
-  const [title, setTitle] = useState('');
-  const [url, setUrl] = useState('');
-  const [nsfw, setNsfw] = useState<boolean>(false);
+  const [text, setText] = useState(postContent ?? '');
+  const [title, setTitle] = useState(postTitle ?? '');
+  const [url, setUrl] = useState(postLink ?? '');
+  const [nsfw, setNsfw] = useState<boolean>(postNsfw ?? false);
   const [languageId, setLanguageId] = useState(
-    userLanguage ?? communityLanguage ?? siteLanguage ?? 0,
+    postLanguage ?? userLanguage ?? communityLanguage ?? siteLanguage ?? 0,
   );
 
   const [selection, setSelection] = useState<ITextSelection>({
@@ -97,7 +114,7 @@ export const useNewPostScreen = (isEdit = false): UseNewPostScreen => {
       }),
     });
 
-    if (draft != null) {
+    if (draft != null && postId == null) {
       setText(draft.content ?? '');
       inputRef.current?.setNativeProps({
         text: draft.content,
@@ -115,10 +132,10 @@ export const useNewPostScreen = (isEdit = false): UseNewPostScreen => {
     });
 
     return unsubscribe;
-  }, [text]);
+  }, [text, title, url, nsfw]);
 
   const beforeRemove = (): void => {
-    if (text !== '' && saveDraft.current) {
+    if (text !== '' && saveDraft.current && postId == null) {
       const newDraft: DraftState = {
         forCommunity: communityId,
         content: text,
@@ -131,6 +148,24 @@ export const useNewPostScreen = (isEdit = false): UseNewPostScreen => {
 
   const onSubmitPress = (): void => {
     submit(async () => {
+      // We have to just edit the post if the post ID is not null
+      if (postId != null) {
+        const res = await instance.editPost({
+          post_id: postId,
+          name: title,
+          body: text !== '' ? text : undefined,
+          url: url !== '' ? url : undefined,
+          nsfw,
+          language_id: languageId,
+        });
+
+        updatePost(res.post_view);
+
+        navigation.pop();
+
+        return;
+      }
+
       const res = await instance.createPost({
         community_id: communityId,
         name: title,
@@ -168,7 +203,10 @@ export const useNewPostScreen = (isEdit = false): UseNewPostScreen => {
     nsfw,
     setNsfw,
 
+    title,
     setTitle,
+
+    url,
     setUrl,
 
     languageId,