+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed typo in checking whether to show the community name in comments
+
 ## [Version 1.0 (46)] - 2023-10-23T11:20:39Z
 
 ### Fixed