+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Toast added to both reply and post screen
 
+### Added
+
+- Added edit comments
+
 ## [Version 1.0 (46)] - 2023-10-23T11:20:39Z
 
 ### Fixed