+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added edit comments
+- Added edit posts
 
 ## [Version 1.0 (46)] - 2023-10-23T11:20:39Z