+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added edit comments
 - Added edit posts
+- Option to show buttons instead of using swipe actions for comments
 
 ## [Version 1.0 (46)] - 2023-10-23T11:20:39Z